requirements.txt
boto3==1.35.82
numpy==1.21.5
pandas==2.2.2
progressbar2==4.5.0
pyarrow==18.1